Fix base href in index.html for GitHub Pages
The web app works locally, but not on GitHub Pages. Changing the base href should fix this.

Additionally, the launchSettings.json file needed to be updated to match the changes to the base href in index.html in order for the change to work locally.

Plus, upload all the code to GitHub Pages (docs/). I'm not sure if this needs to be done every time I change just the index.html file, but I've done it so far...