Fix index.html to work locally when using "dotnet run"
The "Single Page Apps for GitHub Pages" code works for GitHub Pages and when running locally with Visual Studio, but not when using "dotnet run" from the command-line. Eventually we should make this code work for "dotnet run", but for the time being, let's remove it from the local code, but keep it in the code that's used for GitHub Pages (the code under docs/).