Improve UX for inputting data
Before this commit, you needed to either press enter or click the button in order to update the output, but this means that if you select a type from the dropdown, the output wouldn't update. This caused confusion and is pretty annoying. This commit improves the user experience for inputting data by updating the output whenever the data in the input text boxes changes.