Hopefully fix refreshing the page in GitHub Pages site
Refreshing the page in the GitHub Pages site causes a 404 error to be returned. Hopefully this commit will fix this. From the https://github.com/rafrex/spa-github-pages/blob/gh-pages/README.md page:

"Note that if you are setting up a Project Pages site and not using a custom domain (i.e. your site's address is username.github.io/repo-name), then you need to set segmentCount to 1 in the 404.html file in order to keep /repo-name in the path after the redirect."