Improve squad scoring
Imagine the opponent had an grass/electric Pokemon, and one of the Pokemon in your squad was a water Pokemon. Since the grass and electric types are 2x effective against water, the current code sets the defense score to 2+2=4. However, since the opposing Pokemon can only get a max super effectiveness of 2x, the defense score should actually be max(2, 2)=2. The same logic can be applied to the attack score as well.